# docs/swagger.yaml
basePath: /
definitions:
  domain.ArticleInput:
    properties:
      content:
        minLength: 20
        type: string
      title:
        minLength: 10
        type: string
    required:
    - content
    - title
    type: object
  domain.ArticleOutput:
    properties:
      author:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  domain.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  domain.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      surname:
        minLength: 2
        type: string
    required:
    - email
    - name
    - password
    - surname
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Articles Application
  title: Articles App API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      operationId: get-all-articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ArticleOutput'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get All Articles
      tags:
      - Articles
    post:
      consumes:
      - application/json
      operationId: create-articles
      parameters:
      - description: Article input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.ArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create Article
      tags:
      - Articles
  /articles/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete Article
      tags:
      - Articles
    get:
      consumes:
      - application/json
      operationId: get-article-by-id
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ArticleOutput'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get Article By Id
      tags:
      - Articles
    put:
      consumes:
      - application/json
      operationId: update-article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Article input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.ArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update Article
      tags:
      - Articles
  /auth/refresh:
    get:
      consumes:
      - application/json
      operationId: refresh
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
      summary: Refresh
      tags:
      - Users auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      operationId: sign-in
      parameters:
      - description: Sign in input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Sign In
      tags:
      - Users auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      operationId: sign-up
      parameters:
      - description: Sign up input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Sign Up
      tags:
      - Users auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"